fix assignment of timing in voxelization of TG4HitSegments
timing is used to determine where the 'startpoint' and 'endpoint' of the voxels made from true TG4HitSegments.  these are subsequently interpreted as the true begin & end of the energy inside the active volume.
previously, all voxels made from the same TG4HitSegment were assigned the same time, the HitSegment's start time, which meant that for long segments (like those for MIPs), there might be many voxels beyond the start point that would not be interpreted as the end.
fix this to assign intermediate times between TG4HitSegment's start & stop times as appropriate.