handle LEScatter groups w/ no valid parent
this can happen for instance from the following chain:

muon delta ray electron -> brem photon -> compton scatter

The Brem. photon is "invalid" (it's grouped to its parent), but Comptons aren't.  Thus it needs to have its own group.  That case was an edge case that wasn't being handled and now is.