beware integer overflows in ids that come from GEANT trackids...
there are situations where the GEANT trackid is copied into the InstanceID.
because the pileup in the LBNF beam is so large, events (=spills) can routinely have more than 65536 (max of ushort) GEANT tracks in them across the numerous neutrino interactions.
so we need a bigger type to hold the id.